backend/requirements.txt
Flask
Flask-Cors
python-dotenv
Pillow
mysql-connector-python
openai
SQLAlchemy
PyMySQL
Flask-SQLAlchemy
